# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BedemWAF Control API
  version: 0.1.0
  description: |
    REST management API for BedemWAF tenants, applications, origins, policy
    drafts, policy publishing, gateway policy retrieval, and WAF audit event
    search.

    MVP authentication uses static bearer tokens. Administrative `/v1` routes
    use `AdminBearerAuth`; gateway policy retrieval uses `GatewayBearerAuth`.
servers:
  - url: http://localhost:8081
    description: Local development Control API
...
security:
  - AdminBearerAuth: []
paths:
...
  /v1/tenants:
    get:
      tags: [Tenants]
      summary: List tenants
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: object
                required: [tenants]
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Tenants]
      summary: Create tenant
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
            examples:
              demo:
                value:
                  name: Demo Tenant
                  slug: demo
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/apps:
    get:
      tags: [Apps]
      summary: List apps
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Application list
          content:
            application/json:
              schema:
                type: object
                required: [apps]
                properties:
                  apps:
                    type: array
                    items:
                      $ref: "#/components/schemas/App"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Apps]
      summary: Create app with primary origin
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppRequest"
            examples:
              demo:
                value:
                  tenant_id: tenant-1
                  name: Demo App
                  slug: demo-app
                  hostnames: [app.example.local]
                  origin_url: http://localhost:9000
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/apps/{app_id}:
    get:
      tags: [Apps]
      summary: Get app
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppID"
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Apps]
      summary: Update app
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppRequest"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/apps/{app_id}/policies:
    get:
      tags: [Policies]
      summary: List policies for app
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppID"
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                type: object
                required: [policies]
                properties:
                  policies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Policy"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [Policies]
      summary: Create policy draft
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
            examples:
              countPolicy:
                value:
                  name: Default policy
                  mode: count
                  snapshot:
                    mode: count
                    ip_sets:
                      office_ips: [198.51.100.0/24]
                    custom_rules:
                      - id: rule-admin-office-only
                        name: Admin only from office IPs
                        priority: 100
                        enabled: true
                        action: block
                        status_code: 403
                        when:
                          path_starts_with: /admin
                    rate_limits:
                      - id: rl-login
                        name: Login IP limit
                        enabled: true
                        priority: 100
                        key_type: ip
                        limit: 20
                        window_seconds: 60
                        action: block
                        status_code: 429
                    waf:
                      enabled: true
                      engine: coraza
                      rule_engine: DetectionOnly
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/apps/{app_id}/active-policy:
    get:
      tags: [Policies]
      summary: Get active gateway-ready policy for an app
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppID"
      responses:
        "200":
          description: Active gateway-ready policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayPolicy"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/policies/{policy_id}:
    get:
      tags: [Policies]
      summary: Get policy draft metadata and snapshot
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Policies]
      summary: Update mutable policy draft with optimistic locking
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyRequest"
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/policies/{policy_id}/publish:
    post:
      tags: [Policies]
      summary: Publish policy draft to an immutable version
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "201":
          description: Policy published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishPolicyResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/policies/{policy_id}/simulation-summary:
    get:
      tags: [Policies]
      summary: Summarize would-block events for safe rollout
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PolicyID"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
      responses:
        "200":
          description: Policy simulation summary grouped by matching rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicySimulationSummary"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/managed-rule-sets:
    get:
      tags: [Managed Rules]
      summary: List managed rule sets
      security:
        - AdminBearerAuth: []
      responses:
        "200":
          description: Managed rule set list
          content:
            application/json:
              schema:
                type: object
                required: [managed_rule_sets]
                properties:
                  managed_rule_sets:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManagedRuleSet"
        "401":
          $ref: "#/components/responses/Error"
  /v1/managed-rule-sets/{id}/versions:
    get:
      tags: [Managed Rules]
      summary: List versions for a managed rule set
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ManagedRuleSetID"
      responses:
        "200":
          description: Managed rule version list
          content:
            application/json:
              schema:
                type: object
                required: [versions]
                properties:
                  versions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManagedRuleVersion"
        "401":
          $ref: "#/components/responses/Error"
  /v1/managed-rule-sets/{id}/versions/{version_id}/activate:
    post:
      tags: [Managed Rules]
      summary: Placeholder for managed rule version activation
      description: |
        MVP placeholder. The endpoint verifies that the managed rule version
        exists and returns a manual-publish-required status. It does not change
        active gateway policies.
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ManagedRuleSetID"
        - $ref: "#/components/parameters/ManagedRuleVersionID"
      responses:
        "202":
          description: Managed rule version exists; publish a policy manually.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivateManagedRuleVersionResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/gateway/apps/{hostname}/policy:
    get:
      tags: [Gateway]
      summary: Get active gateway policy by hostname
      description: Protected by the separate gateway API key.
      security:
        - GatewayBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Hostname"
      responses:
        "200":
          description: Gateway-ready policy JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayPolicy"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/events/{request_id}:
    get:
      tags: [Events]
      summary: Get WAF audit event by request ID
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200":
          description: Audit event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEvent"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Static MVP admin token from `BEDEMWAF_ADMIN_API_KEY`.
    GatewayBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Static MVP gateway token from `BEDEMWAF_GATEWAY_API_KEY`.
  parameters:
    AppID:
      name: app_id
      in: path
      required: true
      schema:
        type: string
      example: app-1
    PolicyID:
      name: policy_id
      in: path
      required: true
      schema:
        type: string
      example: policy-1
...
  responses:
    Error:
      description: Consistent JSON error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Tenant:
      type: object
      required: [id, name, slug, status, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
        status:
          type: string
          examples: [active]
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...